refactor(utils): move server-only modules to `/app/.server/` directory
